refactor(compat): remove unused per-type SchemaDiff classes (C70/R8)
Delete the legacy per-type diff hierarchy that mirrored the Registry's
instanceof-dispatched typed visitors (SchemaDiffVisitor, ArraySchemaDiff,
ObjectSchemaDiff, NumberSchemaDiff, StringSchemaDiff) plus the zero-reference
JsonSchemaCompatibilityException. The compound-schema diff
(CompoundSchemaDiffVisitor) subsumes them and is strictly more complete.

Fold in a structural cleanup now that the compound visitor is the sole caller
of the schema-compatibility recursion:

- Move isSchemaCompatible, isUnionSchemaCompatible, compareSchema and
  compareSchemaWhenExist out of the generic DiffUtil helper onto
  CompoundSchemaDiffVisitor as private methods (isSchemaCompatible is mutually
  recursive with diffSchemas). DiffUtil is now purely generic diff primitives
  plus small schema accessors.
- Drop diffSchemas to package-private so JsonSchemaCompatibilityChecker is the
  only public compatibility entry point.
- diffSchemas now fails fast (IllegalStateException) rather than silently
  returning when an operand cannot be converted to compound.

Full data-models suite green (1147 tests, 0 failures).
