--- data-models/src/main/java/io/apitomy/datamodels/jsonschema/compat/SchemaDiffVisitor.java
@@ -1,401 +1,0 @@
-package io.apitomy.datamodels.jsonschema.compat;
-
-import io.apitomy.datamodels.models.jsonschema.JFullSchema;
-import io.apitomy.datamodels.models.jsonschema.JsonSchema;
-import io.apitomy.datamodels.models.jsonschema.draft.JDFullSchema;
-import io.apitomy.datamodels.models.jsonschema.draft.draft6.JD6FullSchema;
-import io.apitomy.datamodels.models.jsonschema.draft.draft7.JD7FullSchema;
-import com.fasterxml.jackson.databind.JsonNode;
-
-import java.util.List;
-
-import static io.apitomy.datamodels.jsonschema.compat.DiffType.*;
-import static io.apitomy.datamodels.jsonschema.compat.DiffUtil.*;
-
-/**
- * Top-level diff visitor that dispatches to type-specific visitors based on the schema's type property.
- * <p>
- * In the Registry's everit-based checker, dispatch was via instanceof on typed Schema subclasses.
- * Here we inspect the type property at runtime since the model is flat.
- */
-public class SchemaDiffVisitor {
-
-    private final DiffContext ctx;
-    private final JFullSchema original;
-
-    public SchemaDiffVisitor(DiffContext ctx, JFullSchema original) {
-        this.ctx = ctx;
-        this.original = original;
-    }
-
-    /**
-     * Entry point: compare original and updated schemas.
-     * Any {@code $ref} nodes should be resolved by the dereferencer before
-     * calling this method.
-     */
-    public static void diffSchemas(DiffContext ctx, JFullSchema original, JFullSchema updated) {
-        var pairKey = System.identityHashCode(original)
-                + ":" + System.identityHashCode(updated);
-        if (ctx.visited.contains(pairKey)) {
-            return;
-        }
-        ctx.visited.add(pairKey);
-        try {
-            new SchemaDiffVisitor(ctx, original).visit(updated);
-        } finally {
-            ctx.visited.remove(pairKey);
-        }
-    }
-
-    public void visit(JFullSchema updated) {
-        var originalType = DiffUtil.getTypeString(original);
-        var updatedType = DiffUtil.getTypeString(updated);
-
-        if (originalType != null && updatedType != null && !originalType.equals(updatedType)) {
-            if ("integer".equals(originalType) && "number".equals(updatedType)) {
-                ctx.addDifference(SUBSCHEMA_TYPE_CHANGED_TO_EMPTY_OR_TRUE, originalType, updatedType);
-            } else if (updatedType.isEmpty() || isEmptyOrTrueSchema(updated)) {
-                ctx.addDifference(SUBSCHEMA_TYPE_CHANGED_TO_EMPTY_OR_TRUE, originalType, updatedType);
-            } else {
-                ctx.addDifference(SUBSCHEMA_TYPE_CHANGED, originalType, updatedType);
-            }
-            return;
-        }
-
-        if (originalType != null && updatedType == null) {
-            if (isEmptyOrTrueSchema(updated)) {
-                ctx.addDifference(SUBSCHEMA_TYPE_CHANGED_TO_EMPTY_OR_TRUE, originalType, "");
-                return;
-            }
-            // Check if updated uses composition (anyOf/oneOf) that includes the original type
-            var updAnyOf = updated.getAnyOf();
-            var updOneOf = updated.getOneOf();
-            if (updAnyOf != null || updOneOf != null) {
-                var compositionList = updAnyOf != null ? updAnyOf : updOneOf;
-                // Check if any subschema in the composition is backward-compatible
-                // with the original schema (i.e., the original type is widened into a union)
-                var origMatchesAny = false;
-                for (var sub : compositionList) {
-                    if (sub.isFullSchema()) {
-                        var subCtx = ctx.sub("compositionCheck");
-                        if (isSchemaCompatible(subCtx, original, sub.asFullSchema(), true)) {
-                            origMatchesAny = true;
-                            break;
-                        }
-                    }
-                }
-                if (origMatchesAny) {
-                    ctx.addDifference(SUBSCHEMA_TYPE_CHANGED_TO_EMPTY_OR_TRUE, originalType, "anyOf/oneOf");
-                    return;
-                }
-            }
-        }
-
-        if (originalType == null && updatedType != null) {
-            if (isEmptyOrTrueSchema(original)) {
-                ctx.addDifference(SUBSCHEMA_TYPE_CHANGED, "", updatedType);
-                return;
-            }
-        }
-
-        var effectiveType = originalType != null ? originalType : updatedType;
-        if (effectiveType == null) {
-            var origTypeList = DiffUtil.getTypeList(original);
-            var updTypeList = DiffUtil.getTypeList(updated);
-            if ((origTypeList != null && origTypeList.size() > 1)
-                    || (updTypeList != null && updTypeList.size() > 1)) {
-                ctx.addUnsupported("Multi-valued type field (e.g. type: [\"string\", \"number\"])");
-            }
-            diffAllPropertyGroups(original, updated);
-            return;
-        }
-
-        switch (effectiveType) {
-            case "object" -> new ObjectSchemaDiff(ctx, original, updated).visit();
-            case "array" -> new ArraySchemaDiff(ctx, original, updated).visit();
-            case "string" -> new StringSchemaDiff(ctx, original, updated).visit();
-            case "number", "integer" -> new NumberSchemaDiff(ctx, original, updated, effectiveType).visit();
-            case "boolean" -> { /* no type-specific constraints */ }
-            case "null" -> { /* no type-specific constraints */ }
-            default -> { /* unknown type, nothing to compare */ }
-        }
-
-        diffCompositionKeywords(original, updated);
-        diffConditionalKeywords(original, updated);
-        diffEnumConst(original, updated);
-        diffReference(original, updated);
-    }
-
-    private boolean isEmptyOrTrueSchema(JFullSchema schema) {
-        if (DiffUtil.getTypeString(schema) != null
-                || schema.getAllOf() != null
-                || schema.getAnyOf() != null
-                || schema.getOneOf() != null
-                || schema.getNot() != null
-                || schema.getEnum() != null
-                || schema.getProperties() != null
-                || schema.getRequired() != null
-                || schema.getMinLength() != null
-                || schema.getMaxLength() != null
-                || schema.getMinItems() != null
-                || schema.getMaxItems() != null
-                || schema.getMinProperties() != null
-                || schema.getMaxProperties() != null
-                || schema.getPattern() != null
-                || schema.getFormat() != null
-                || schema.getMultipleOf() != null
-                || schema.getAdditionalProperties() != null
-                || schema.getPatternProperties() != null
-                || getConst(schema) != null) {
-            return false;
-        }
-        if (schema instanceof JDFullSchema d
-                && (d.getMinimum() != null || d.getMaximum() != null
-                    || d.getItems() != null || d.getAdditionalItems() != null)) {
-            return false;
-        }
-        return true;
-    }
-
-    private void diffAllPropertyGroups(JFullSchema original, JFullSchema updated) {
-        new ObjectSchemaDiff(ctx, original, updated).visit();
-        new ArraySchemaDiff(ctx, original, updated).visit();
-        new StringSchemaDiff(ctx, original, updated).visit();
-        new NumberSchemaDiff(ctx, original, updated, null).visit();
-        diffCompositionKeywords(original, updated);
-        diffConditionalKeywords(original, updated);
-        diffEnumConst(original, updated);
-        diffReference(original, updated);
-    }
-
-    private void diffCompositionKeywords(JFullSchema original, JFullSchema updated) {
-        var origAllOf = original.getAllOf();
-        var origAnyOf = original.getAnyOf();
-        var origOneOf = original.getOneOf();
-        var updAllOf = updated.getAllOf();
-        var updAnyOf = updated.getAnyOf();
-        var updOneOf = updated.getOneOf();
-
-        // allOf → anyOf is backward compatible (widening: from "must satisfy ALL" to "must satisfy ANY")
-        // oneOf → anyOf is backward compatible (widening: from "must satisfy EXACTLY ONE" to "must satisfy ANY")
-        if (origAllOf != null && updAnyOf != null && updAllOf == null) {
-            diffCompositionList(ctx, origAllOf, updAnyOf,
-                    COMBINED_TYPE_CRITERION_EXTENDED, COMBINED_TYPE_CRITERION_EXTENDED);
-            ctx.addDifference(COMBINED_TYPE_CRITERION_EXTENDED, "allOf", "anyOf");
-        } else if (origOneOf != null && updAnyOf != null && updOneOf == null) {
-            diffCompositionList(ctx, origOneOf, updAnyOf,
-                    COMBINED_TYPE_CRITERION_EXTENDED, COMBINED_TYPE_CRITERION_EXTENDED);
-            ctx.addDifference(COMBINED_TYPE_CRITERION_EXTENDED, "oneOf", "anyOf");
-        } else if (origAnyOf != null && updAllOf != null && updAnyOf == null) {
-            ctx.addDifference(COMBINED_TYPE_CRITERION_NARROWED, "anyOf", "allOf");
-        } else if (origAnyOf != null && updOneOf != null && updAnyOf == null) {
-            ctx.addDifference(COMBINED_TYPE_CRITERION_NARROWED, "anyOf", "oneOf");
-        } else {
-            diffCompositionList(ctx, origAllOf, updAllOf,
-                    COMBINED_TYPE_ALL_OF_SIZE_INCREASED, COMBINED_TYPE_ALL_OF_SIZE_DECREASED);
-            diffCompositionList(ctx, origAnyOf, updAnyOf,
-                    COMBINED_TYPE_ANY_OF_SIZE_INCREASED, COMBINED_TYPE_ANY_OF_SIZE_DECREASED);
-            diffCompositionList(ctx, origOneOf, updOneOf,
-                    COMBINED_TYPE_ONE_OF_SIZE_INCREASED, COMBINED_TYPE_ONE_OF_SIZE_DECREASED);
-        }
-
-        diffNotSchema(original, updated);
-    }
-
-    private void diffCompositionList(DiffContext ctx,
-                                     List<JsonSchema> originalList,
-                                     List<JsonSchema> updatedList,
-                                     DiffType increasedType, DiffType decreasedType) {
-        if (originalList == null && updatedList == null) return;
-        if (originalList == null || updatedList == null) {
-            ctx.addDifference(COMBINED_TYPE_CRITERION_CHANGED, originalList, updatedList);
-            return;
-        }
-
-        if (updatedList.size() > originalList.size()) {
-            ctx.addDifference(increasedType, originalList.size(), updatedList.size());
-        } else if (updatedList.size() < originalList.size()) {
-            ctx.addDifference(decreasedType, originalList.size(), updatedList.size());
-        }
-
-        var unmatchedCount = 0;
-        for (var updSub : updatedList) {
-            var matched = false;
-            for (var origSub : originalList) {
-                var subCtx = ctx.sub("composition");
-                if (isUnionSchemaCompatible(subCtx, origSub, updSub, true)
-                        && isUnionSchemaCompatible(subCtx, origSub, updSub, false)) {
-                    matched = true;
-                    break;
-                }
-            }
-            if (!matched) {
-                for (var origSub : originalList) {
-                    var subCtx = ctx.sub("composition");
-                    if (isUnionSchemaCompatible(subCtx, origSub, updSub, true)) {
-                        matched = true;
-                        break;
-                    }
-                }
-            }
-            if (!matched) {
-                unmatchedCount++;
-            }
-        }
-        var newSubschemas = Math.max(0, updatedList.size() - originalList.size());
-        var changedSubschemas = unmatchedCount - newSubschemas;
-        if (changedSubschemas > 0) {
-            ctx.addDifference(COMBINED_TYPE_SUBSCHEMA_NOT_COMPATIBLE, originalList, updatedList);
-        }
-    }
-
-    private void diffNotSchema(JFullSchema original, JFullSchema updated) {
-        var origNot = original.getNot();
-        var updNot = updated.getNot();
-        if (origNot == null && updNot == null) return;
-
-        compareSchema(ctx, origNot, updNot,
-                // not schema: backwards compatibility is inverted
-                SUBSCHEMA_TYPE_CHANGED, SUBSCHEMA_TYPE_CHANGED,
-                NOT_TYPE_SCHEMA_COMPATIBLE_BOTH,
-                NOT_TYPE_SCHEMA_COMPATIBLE_BACKWARD_NOT_FORWARD,
-                NOT_TYPE_SCHEMA_COMPATIBLE_FORWARD_NOT_BACKWARD,
-                NOT_TYPE_SCHEMA_COMPATIBLE_NONE);
-    }
-
-    private void diffConditionalKeywords(JFullSchema original, JFullSchema updated) {
-        var origIf = getConditional(original, "if");
-        var updIf = getConditional(updated, "if");
-        var origThen = getConditional(original, "then");
-        var updThen = getConditional(updated, "then");
-        var origElse = getConditional(original, "else");
-        var updElse = getConditional(updated, "else");
-
-        if (origIf == null && updIf == null && origThen == null && updThen == null
-                && origElse == null && updElse == null) {
-            return;
-        }
-
-        compareSchema(ctx, origIf, updIf,
-                CONDITIONAL_TYPE_IF_SCHEMA_ADDED, CONDITIONAL_TYPE_IF_SCHEMA_REMOVED,
-                CONDITIONAL_TYPE_IF_SCHEMA_COMPATIBLE_BOTH,
-                CONDITIONAL_TYPE_IF_SCHEMA_COMPATIBLE_BACKWARD_NOT_FORWARD,
-                CONDITIONAL_TYPE_IF_SCHEMA_COMPATIBLE_FORWARD_NOT_BACKWARD,
-                CONDITIONAL_TYPE_IF_SCHEMA_COMPATIBLE_NONE);
-
-        compareSchema(ctx, origThen, updThen,
-                CONDITIONAL_TYPE_THEN_SCHEMA_ADDED, CONDITIONAL_TYPE_THEN_SCHEMA_REMOVED,
-                CONDITIONAL_TYPE_THEN_SCHEMA_COMPATIBLE_BOTH,
-                CONDITIONAL_TYPE_THEN_SCHEMA_COMPATIBLE_BACKWARD_NOT_FORWARD,
-                CONDITIONAL_TYPE_THEN_SCHEMA_COMPATIBLE_FORWARD_NOT_BACKWARD,
-                CONDITIONAL_TYPE_THEN_SCHEMA_COMPATIBLE_NONE);
-
-        compareSchema(ctx, origElse, updElse,
-                CONDITIONAL_TYPE_ELSE_SCHEMA_ADDED, CONDITIONAL_TYPE_ELSE_SCHEMA_REMOVED,
-                CONDITIONAL_TYPE_ELSE_SCHEMA_COMPATIBLE_BOTH,
-                CONDITIONAL_TYPE_ELSE_SCHEMA_COMPATIBLE_BACKWARD_NOT_FORWARD,
-                CONDITIONAL_TYPE_ELSE_SCHEMA_COMPATIBLE_FORWARD_NOT_BACKWARD,
-                CONDITIONAL_TYPE_ELSE_SCHEMA_COMPATIBLE_NONE);
-    }
-
-    private static JsonSchema getConditional(JFullSchema schema, String keyword) {
-        return switch (keyword) {
-            case "if" -> {
-                if (schema instanceof JD7FullSchema d) yield d.getIf();
-                yield null;
-            }
-            case "then" -> {
-                if (schema instanceof JD7FullSchema d) yield d.getThen();
-                yield null;
-            }
-            case "else" -> {
-                if (schema instanceof JD7FullSchema d) yield d.getElse();
-                yield null;
-            }
-            default -> null;
-        };
-    }
-
-    private void diffEnumConst(JFullSchema original, JFullSchema updated) {
-        var origEnum = original.getEnum();
-        var updEnum = updated.getEnum();
-        var origConst = getConst(original);
-        var updConst = getConst(updated);
-
-        if (origEnum != null || updEnum != null) {
-            if (origEnum == null) {
-                if (origConst != null && updEnum.size() == 1
-                        && updEnum.get(0).toString().equals(origConst.toString())) {
-                    // enum added is equivalent to existing const — no diff
-                } else {
-                    ctx.addDifference(ENUM_TYPE_VALUES_ADDED, null, updEnum);
-                }
-            } else if (updEnum == null) {
-                if (updConst != null && origEnum.size() == 1
-                        && origEnum.get(0).toString().equals(updConst.toString())) {
-                    // enum removed is equivalent to new const — no diff
-                } else {
-                    ctx.addDifference(ENUM_TYPE_VALUES_CHANGED, origEnum, null);
-                }
-            } else {
-                var origSet = new java.util.HashSet<>(origEnum.stream().map(Object::toString).toList());
-                var updSet = new java.util.HashSet<>(updEnum.stream().map(Object::toString).toList());
-                if (!origSet.equals(updSet)) {
-                    ctx.addDifference(ENUM_TYPE_VALUES_CHANGED, origEnum, updEnum);
-                    for (var v : updSet) {
-                        if (!origSet.contains(v)) {
-                            ctx.addDifference(ENUM_TYPE_VALUES_MEMBER_ADDED, null, v);
-                        }
-                    }
-                    for (var v : origSet) {
-                        if (!updSet.contains(v)) {
-                            ctx.addDifference(ENUM_TYPE_VALUES_MEMBER_REMOVED, v, null);
-                        }
-                    }
-                }
-            }
-        }
-
-        diffConst(original, updated);
-    }
-
-    private void diffConst(JFullSchema original, JFullSchema updated) {
-        var origConst = getConst(original);
-        var updConst = getConst(updated);
-        if (origConst == null && updConst == null) return;
-        if (origConst != null && updConst != null) {
-            if (!origConst.equals(updConst)) {
-                ctx.addDifference(CONST_TYPE_VALUE_CHANGED, origConst, updConst);
-            }
-        } else if (origConst == null) {
-            var origEnum = original.getEnum();
-            if (origEnum != null && origEnum.size() == 1
-                    && origEnum.get(0).toString().equals(updConst.toString())) {
-                return;
-            }
-            ctx.addDifference(CONST_TYPE_VALUE_ADDED, null, updConst);
-        } else {
-            var updEnum = updated.getEnum();
-            if (updEnum != null && updEnum.size() == 1
-                    && updEnum.get(0).toString().equals(origConst.toString())) {
-                return;
-            }
-            ctx.addDifference(CONST_TYPE_VALUE_REMOVED, origConst, null);
-        }
-    }
-
-    private static JsonNode getConst(JFullSchema schema) {
-        if (schema instanceof JD6FullSchema d) return d.getConst();
-        if (schema instanceof JD7FullSchema d) return d.getConst();
-        return null;
-    }
-
-    private void diffReference(JFullSchema original, JFullSchema updated) {
-        var origRef = DiffUtil.get$ref(original);
-        var updRef = DiffUtil.get$ref(updated);
-
-        if (origRef != null && updRef != null && !origRef.equals(updRef)) {
-            ctx.addDifference(REFERENCE_TYPE_TARGET_SCHEMA_CHANGED, origRef, updRef);
-        }
-    }
-}
